fix(KvRmsNormRopeCache): 量化 rope GM 偏移 uint32->int64 修复溢出回绕



# message auto-generated for no-merge-commit merge:
!7656 merge fix/kvrmsnormropecache-rope-offset-int64 into master

fix(KvRmsNormRopeCache): 量化 rope GM 偏移 uint32->int64 修复溢出回绕

Created-by: qianzehong
Commit-by: qianzehong
Merged-by: cann-robot
Description: ## 关联 Issue
#3351

## 根因
regbase 量化 rope 路径（`op_kernel/arch35/kv_rms_norm_rope_cache_regbase_recompute.h`）的 GM 元素偏移用了 `uint32_t`，`rowIdx*dk` 超过 2^32 时回绕成错误地址 → VEC 访问 UB 越界（aivec 341 / 507035）。非量化路径 `RopeWithoutQuant` 与 dv 侧偏移本就是 `int64_t`，不受影响。

## 修改
- `kv_rms_norm_rope_cache_regbase_recompute.h`：`RopeSymQuant` / `RopeAsymQuant` / `RopeSymQuantWithKCache` / `RopeAsymQuantWithKCache` 中 `cosSinOffset1/2`、`kScaleOffset1/2`、`kOffset1/2`、`kOutOffset1/2` 由 `uint32_t` 改为 `int64_t`（对齐非量化路径与 dv 侧）。不动 full_load（仅小 D 可命中，不会溢出）。

## 验证
- 现有 op_host / op_api UT 全过（44 + 9 PASSED），op_kernel UT 编译通过，无回归。
- Ascend950 上板复现/确认：修复前 `B=3, dk=2^31` 崩溃（`507035 / aivec 341`）；改 int64 后同用例不再崩溃、cache 各行结果正确（0.5/1.0/2.0）。
- 注：触发需 2^32 元素，CPU 仿真 UT 无法分配，复现/验证在真实 NPU（~60GiB HBM）上完成。



See merge request: cann/ops-transformer!7656

+++ posembedding/kv_rms_norm_rope_cache/op_kernel/arch35/kv_rms_norm_rope_cache_regbase_recompute.h
@@ -1491,13 +1491,13 @@ public:
             }
             
             int64_t kInOffset = xDimOffset + this->dv + this->ubFactor * ubIdx;
-            uint32_t cosSinOffset1 = cosSinDimOffset + this->ubFactor * ubIdx / CONST_TWO;
-            uint32_t cosSinOffset2 = this->dk / CONST_TWO + cosSinOffset1;
-            uint32_t kScaleOffset1 = this->ubFactor * ubIdx / CONST_TWO;
-            uint32_t kScaleOffset2 = this->dk / CONST_TWO + kScaleOffset1;
+            int64_t cosSinOffset1 = cosSinDimOffset + this->ubFactor * ubIdx / CONST_TWO;
+            int64_t cosSinOffset2 = this->dk / CONST_TWO + cosSinOffset1;
+            int64_t kScaleOffset1 = this->ubFactor * ubIdx / CONST_TWO;
+            int64_t kScaleOffset2 = this->dk / CONST_TWO + kScaleOffset1;
             // 中间结果
-            uint32_t kOutOffset1 = rowIdx * this->dk + this->ubFactor * ubIdx / CONST_TWO;
-            uint32_t kOutOffset2 = this->dk / CONST_TWO + kOutOffset1;
+            int64_t kOutOffset1 = rowIdx * this->dk + this->ubFactor * ubIdx / CONST_TWO;
+            int64_t kOutOffset2 = this->dk / CONST_TWO + kOutOffset1;
 
             uint32_t cosSinLen = tmpFactor / CONST_TWO;
             uint32_t alignOffset = this->ubFactor / CONST_TWO;
@@ -1713,12 +1713,12 @@ public:
             }
             
             int64_t kInOffset = xDimOffset + this->dv + this->ubFactor * ubIdx;
-            uint32_t cosSinOffset1 = cosSinDimOffset + this->ubFactor * ubIdx / CONST_TWO;
-            uint32_t cosSinOffset2 = this->dk / CONST_TWO + cosSinOffset1;
-            uint32_t kScaleOffset1 = this->ubFactor * ubIdx / CONST_TWO;
-            uint32_t kScaleOffset2 = this->dk / CONST_TWO + kScaleOffset1;
-            uint32_t kOffset1 = this->ubFactor * ubIdx / CONST_TWO;
-            uint32_t kOffset2 = this->dk / CONST_TWO + kOffset1;
+            int64_t cosSinOffset1 = cosSinDimOffset + this->ubFactor * ubIdx / CONST_TWO;
+            int64_t cosSinOffset2 = this->dk / CONST_TWO + cosSinOffset1;
+            int64_t kScaleOffset1 = this->ubFactor * ubIdx / CONST_TWO;
+            int64_t kScaleOffset2 = this->dk / CONST_TWO + kScaleOffset1;
+            int64_t kOffset1 = this->ubFactor * ubIdx / CONST_TWO;
+            int64_t kOffset2 = this->dk / CONST_TWO + kOffset1;
 
             uint32_t cosSinLen = tmpFactor / CONST_TWO;
             uint32_t alignOffset = this->ubFactor / CONST_TWO;
@@ -1939,10 +1939,10 @@ public:
             }
             
             int64_t kInOffset = xDimOffset + this->dv + this->ubFactor * ubIdx;
-            uint32_t cosSinOffset1 = cosSinDimOffset + this->ubFactor * ubIdx / CONST_TWO;
-            uint32_t cosSinOffset2 = this->dk / CONST_TWO + cosSinOffset1;
-            uint32_t kScaleOffset1 = this->ubFactor * ubIdx / CONST_TWO;
-            uint32_t kScaleOffset2 = this->dk / CONST_TWO + kScaleOffset1;
+            int64_t cosSinOffset1 = cosSinDimOffset + this->ubFactor * ubIdx / CONST_TWO;
+            int64_t cosSinOffset2 = this->dk / CONST_TWO + cosSinOffset1;
+            int64_t kScaleOffset1 = this->ubFactor * ubIdx / CONST_TWO;
+            int64_t kScaleOffset2 = this->dk / CONST_TWO + kScaleOffset1;
 
             uint32_t cosSinLen = tmpFactor / CONST_TWO;
             uint32_t alignOffset = this->ubFactor / CONST_TWO;
@@ -2236,15 +2236,15 @@ public:
             }
             
             int64_t kInOffset = xDimOffset + this->dv + this->ubFactor * ubIdx;
-            uint32_t cosSinOffset1 = cosSinDimOffset + this->ubFactor * ubIdx / CONST_TWO;
-            uint32_t cosSinOffset2 = this->dk / CONST_TWO + cosSinOffset1;
-            uint32_t kScaleOffset1 = this->ubFactor * ubIdx / CONST_TWO;
-            uint32_t kScaleOffset2 = this->dk / CONST_TWO + kScaleOffset1;
-            uint32_t kOffset1 = this->ubFactor * ubIdx / CONST_TWO;
-            uint32_t kOffset2 = this->dk / CONST_TWO + kOffset1;
+            int64_t cosSinOffset1 = cosSinDimOffset + this->ubFactor * ubIdx / CONST_TWO;
+            int64_t cosSinOffset2 = this->dk / CONST_TWO + cosSinOffset1;
+            int64_t kScaleOffset1 = this->ubFactor * ubIdx / CONST_TWO;
+            int64_t kScaleOffset2 = this->dk / CONST_TWO + kScaleOffset1;
+            int64_t kOffset1 = this->ubFactor * ubIdx / CONST_TWO;
+            int64_t kOffset2 = this->dk / CONST_TWO + kOffset1;
             // 中间结果
-            uint32_t kOutOffset1 = rowIdx * this->dk + this->ubFactor * ubIdx / CONST_TWO;
-            uint32_t kOutOffset2 = this->dk / CONST_TWO + kOutOffset1;
+            int64_t kOutOffset1 = rowIdx * this->dk + this->ubFactor * ubIdx / CONST_TWO;
+            int64_t kOutOffset2 = this->dk / CONST_TWO + kOutOffset1;
 
             uint32_t cosSinLen = tmpFactor / CONST_TWO;
             uint32_t alignOffset = this->ubFactor / CONST_TWO;
